WiP on gen_metadata fixing. README.md updated
