+++ README.md
@@ -26,4 +26,8 @@ Plagiarism identifier
 	- cd data/skip-thoughts
 	- wget http://www.cs.toronto.edu/~rkiros/models/uni_skip.npz http://www.cs.toronto.edu/~rkiros/models/uni_skip.npz.pkl http://www.cs.toronto.edu/~rkiros/models/bi_skip.npz http://www.cs.toronto.edu/~rkiros/models/bi_skip.npz.pk
 	- cd ../..
-	- python ./src/gen_stvecs_for_docs.py --pandb plag_train.db --stdir ./data/skip-thoughts --destdir ./data --vocab ./data/skip-thoughts/dictionary.txt --start 1
+	- python ./src/create_st_model.py --pandb plag_train.db --stdir ./data/skip-thoughts --destdir ./data --vocab ./data/skip-thoughts/dictionary.txt --start 1
+
+### Gerar metadados:
+  - pip install numpy scipy matplotlib ipython jupyter pandas sympy nose 
+  - python ./src/gen_metadata.py
